backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.72.0
anyascii==0.3.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncer==0.0.10
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.1.0
backoff==2.2.1
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chainlit==2.8.4
chardet==5.2.0
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.6
cssselect==1.3.0
cssutils==2.11.1
cuid==0.4
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
duckduckgo_search==6.3.2
easyocr==1.7.2
ecdsa==0.19.1
einops==0.8.1
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastapi==0.116.2
filelock==3.18.0
filetype==1.2.0
Flask==3.0.3
flask-cors==6.0.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.10
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.36.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.8
Lazify==0.4.0
lazy_loader==0.4
librosa==0.11.0
lime==0.2.0.1
literalai==0.1.201
llvmlite==0.45.1
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
mcp==1.20.0
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.1
numba==0.62.1
numpy==2.2.6
ollama==0.6.0
onnx==1.18.0
openai==1.97.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-alephalpha==0.47.5
opentelemetry-instrumentation-anthropic==0.47.5
opentelemetry-instrumentation-bedrock==0.47.5
opentelemetry-instrumentation-chromadb==0.47.5
opentelemetry-instrumentation-cohere==0.47.5
opentelemetry-instrumentation-crewai==0.47.5
opentelemetry-instrumentation-google-generativeai==0.47.5
opentelemetry-instrumentation-groq==0.47.5
opentelemetry-instrumentation-haystack==0.47.5
opentelemetry-instrumentation-lancedb==0.47.5
opentelemetry-instrumentation-langchain==0.47.5
opentelemetry-instrumentation-llamaindex==0.47.5
opentelemetry-instrumentation-logging==0.59b0
opentelemetry-instrumentation-marqo==0.47.5
opentelemetry-instrumentation-mcp==0.47.5
opentelemetry-instrumentation-milvus==0.47.5
opentelemetry-instrumentation-mistralai==0.47.5
opentelemetry-instrumentation-ollama==0.47.5
opentelemetry-instrumentation-openai==0.47.5
opentelemetry-instrumentation-openai-agents==0.47.5
opentelemetry-instrumentation-pinecone==0.47.5
opentelemetry-instrumentation-qdrant==0.47.5
opentelemetry-instrumentation-redis==0.59b0
opentelemetry-instrumentation-replicate==0.47.5
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-sagemaker==0.47.5
opentelemetry-instrumentation-sqlalchemy==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-instrumentation-together==0.47.5
opentelemetry-instrumentation-transformers==0.47.5
opentelemetry-instrumentation-urllib3==0.59b0
opentelemetry-instrumentation-vertexai==0.47.5
opentelemetry-instrumentation-watsonx==0.47.5
opentelemetry-instrumentation-weaviate==0.47.5
opentelemetry-instrumentation-writer==0.47.5
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.59b0
opt-einsum==3.3.0
orjson==3.11.0
outcome==1.3.0.post0
packaging==25.0
paddleocr==3.1.0
paddlepaddle==3.1.0
paddlex==3.1.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.3.0
platformdirs==4.7.0
pooch==1.9.0
posthog==3.25.0
premailer==3.10.0
prettytable==3.16.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
pymongo==4.13.2
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-doctr==1.0.0
python-dotenv==1.1.1
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.19
python-socketio==5.14.3
pytz==2025.2
pywin32==311
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.8.0
scipy==1.16.0
seaborn==0.13.2
selenium==4.35.0
sentence-transformers==2.7.0
setuptools==80.9.0
shap==0.49.1
shapely==2.1.1
simple-websocket==1.1.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==1.0.0
SQLAlchemy==2.0.41
sse-starlette==3.0.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
torch==2.10.0
tornado==6.5.2
tqdm==4.67.1
traceloop-sdk==0.47.5
transformers==4.57.6
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
validators==0.35.0
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
wsproto==1.2.0
xgboost==3.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
